+++ worlds/chibi_robo/CHANGELOG.md
@@ -1,4 +1,12 @@
 # Chibi Robo - Changelog
+## v1.1.3 - Fixed Locations / Code Clean up
+- Fixed Some Locations Not Sending
+- Ship is no longer in the item / location pool due to player not able to pick up any other item at that location but may add back later.
+- 'Progressive' or 'Useful' items will show up as Cookies in world while any other classification shows up a Wastepaper to help with players only being able to hold 10 of the same item at once.
+
+## v1.1.2 - Fix UFOs Flag Softlock
+- Removed Backyard Weeds Locations
+
 ## v1.1.1 - Add Download Link Via Web
 - Add support to download required patch file from the web download links in room so host doesn't need to send each chibi robo player their patch file